+++ StampCollection/AlbumType+Ex.swift
@@ -102,7 +102,7 @@ extension AlbumType {
     var theTotalPrice: String {
         // will return the total price of theRefs array
         let total = theFamilies.reduce(0.0, {
-            $0 + ($1.theTotalPrice.toFloat() ?? 0.0)
+            $0 + ($1.theTotalPrice.toDouble() ?? 0.0)
         })
         return String(format: "%.2f", total)
     }
